backend/requirements.txt
# LiveKit Agent Dependencies
livekit>=0.17.0
livekit-agents>=0.9.0
livekit-plugins-openai>=0.7.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-elevenlabs>=0.6.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0

# AI/ML
openai>=1.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
pydantic>=2.0.0

# Image processing
scikit-image>=0.21.0

# Logging
loguru>=0.7.0
